+++ tests/rewards/apply_operation_test.py
@@ -1,12 +1,14 @@
+from rlgym_learn_logging.api.rewards.log import SEPARATOR
+from rlgym_learn_logging.api.wrappers.apply_operation_wrapper import (
+    ApplyOperationWrapper,
+)
+from rlgym_learn_logging.logging_utils import REWARDS_HEADER
 from tests.create_env import (
     ConstantReward,
     RLGymUnittestCase,
     create_env,
     generate_random_actions,
 )
-from void_logging.api.rewards.log import SEPARATOR
-from void_logging.api.wrappers.apply_operation_wrapper import ApplyOperationWrapper
-from void_logging.logging_utils import REWARDS_HEADER
 
 
 class ApplyOperationWrapperTestCase(RLGymUnittestCase):